LayoutTests/media/track/track-cue-rendering-snap-to-lines-not-set-expected.txt


Test that TextTrack's cues are rendered correctly when the snap to lines flag is not set.
EVENT(canplaythrough)

** Default **
EVENT(seeked)
Active cues: 1

* Cue 0 *
EXPECTED (cueDisplayElement.innerText == 'This is a demo track to illustrate positioning features of cues.') OK
Cue alignment (default or as specified): middle
Cue left position (percentage): 50
Cue top position (percentage): 100

** Center **
EVENT(seeked)
Active cues: 1

* Cue 0 *
EXPECTED (cueDisplayElement.innerText == 'Using HTML5 captions, you can position the cues anywhere.') OK
Cue alignment (default or as specified): middle
Cue left position (percentage): 50
Cue top position (percentage): 50

** Top left **
EVENT(seeked)
Active cues: 1

* Cue 0 *
EXPECTED (cueDisplayElement.innerText == 'For example, here is a cue on the top left-corner.') OK
Cue alignment (default or as specified): start
Cue left position (percentage): 0
Cue top position (percentage): 0

** Top right **
EVENT(seeked)
Active cues: 1

* Cue 0 *
EXPECTED (cueDisplayElement.innerText == 'Or on the top-right corner.') OK
Cue alignment (default or as specified): end
Cue left position (percentage): 100
Cue top position (percentage): 0

** Bottom right **
EVENT(seeked)
Active cues: 1

* Cue 0 *
EXPECTED (cueDisplayElement.innerText == 'Bottom-right corner.') OK
Cue alignment (default or as specified): right
Cue left position (percentage): 100
Cue top position (percentage): 100

** Bottom left **
EVENT(seeked)
Active cues: 1

* Cue 0 *
EXPECTED (cueDisplayElement.innerText == 'And, as you might expect, the bottom-left one.') OK
Cue alignment (default or as specified): left
Cue left position (percentage): 0
Cue top position (percentage): 100

** Bottom center **
EVENT(seeked)
Active cues: 1

* Cue 0 *
EXPECTED (cueDisplayElement.innerText == 'Cues are implicitly displayed within the viewport area.') OK
Cue alignment (default or as specified): middle
Cue left position (percentage): 50
Cue top position (percentage): 50

** Multiple cues **
EVENT(seeked)
Active cues: 7

* Cue 0 *
EXPECTED (cueDisplayElement.innerText == 'You may even have multiple cues in the same time.') OK
Cue alignment (default or as specified): middle
Cue left position (percentage): 50
Cue top position (percentage): 50

* Cue 1 *
EXPECTED (cueDisplayElement.innerText == 'Top-left cue.') OK
Cue alignment (default or as specified): start
Cue left position (percentage): 0
Cue top position (percentage): 0

* Cue 2 *
EXPECTED (cueDisplayElement.innerText == 'Top-right cue.') OK
Cue alignment (default or as specified): end
Cue left position (percentage): 100
Cue top position (percentage): 7

* Cue 3 *
EXPECTED (cueDisplayElement.innerText == 'Bottom cue.') OK
Cue alignment (default or as specified): middle
Cue left position (percentage): 50
Cue top position (percentage): 100

* Cue 4 *
EXPECTED (cueDisplayElement.innerText == 'Random cue 1.') OK
Cue alignment (default or as specified): end
Cue left position (percentage): 100
Cue top position (percentage): 25

* Cue 5 *
EXPECTED (cueDisplayElement.innerText == 'Random cue 2.') OK
Cue alignment (default or as specified): start
Cue left position (percentage): 44
Cue top position (percentage): 33

* Cue 6 *
EXPECTED (cueDisplayElement.innerText == 'Random cue 3.') OK
Cue alignment (default or as specified): start
Cue left position (percentage): 23
Cue top position (percentage): 43
END OF TEST
